backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
python-dotenv
requests
ffmpeg-python
strands-agents
strands-agents-tools
strands-agents-builder
aiofiles
python-dateutil
